Put F back in to the model
The population density, F, was removed from the model because it is not necessary.  It is used so that the overall mean can be computed.  It is not strictly necessary to do this, but it is nice to use the overall mean instead of the referent mean.  We need F in the model so that we can combine ITHIM objects to create new ITHIM objects.  The value for meanType can still be set to referent if F is not available.
close #6